refac: Moved entry point to higher level (#69)
* Updates

* Update test_electrical_heating_job_args.py

* Updates

* Update execute_with_deps.py

* Updates
