+++ source/capacity_settlement/src/capacity_settlement/application/execute_with_deps.py
@@ -1,5 +1,4 @@
-﻿import os
-import sys
+﻿import sys
 from argparse import Namespace
 from collections.abc import Callable
 
@@ -7,26 +6,16 @@ import telemetry_logging.logging_configuration as config
 from opentelemetry.trace import SpanKind
 from telemetry_logging.span_recording import span_record_exception
 
-from source.capacity_settlement.src.capacity_settlement.application.entry_points.job_args.capacity_settlement_args import (
+from source.capacity_settlement.src.capacity_settlement.application.job_args.capacity_settlement_args import (
     CapacitySettlementArgs,
 )
-from source.capacity_settlement.src.capacity_settlement.application.entry_points.job_args.capacity_settlement_job_args import (
+from source.capacity_settlement.src.capacity_settlement.application.job_args.capacity_settlement_job_args import (
     parse_command_line_arguments,
     parse_job_arguments,
 )
 
 
-def execute() -> None:
-    applicationinsights_connection_string = os.getenv(
-        "APPLICATIONINSIGHTS_CONNECTION_STRING"
-    )
-
-    start_with_deps(
-        applicationinsights_connection_string=applicationinsights_connection_string,
-    )
-
-
-def start_with_deps(
+def execute_with_deps(
     *,
     cloud_role_name: str = "dbr-capacity-settlement",
     applicationinsights_connection_string: str | None = None,
