+++ source/capacity_settlement/src/capacity_settlement/entry_point.py
@@ -1,0 +1,15 @@
+﻿import os
+
+from source.capacity_settlement.src.capacity_settlement.application.execute_with_deps import (
+    execute_with_deps,
+)
+
+
+def execute() -> None:
+    applicationinsights_connection_string = os.getenv(
+        "APPLICATIONINSIGHTS_CONNECTION_STRING"
+    )
+
+    execute_with_deps(
+        applicationinsights_connection_string=applicationinsights_connection_string,
+    )
