+++ source/capacity_settlement/tests/capacity_settlement_tests/application/job_args/test_capacity_settlement_job_args.py
@@ -2,11 +2,11 @@
 
 import pytest
 
-from source.electrical_heating.src.electrical_heating.application.entry_points.job_args.electrical_heating_job_args import (
-    parse_job_arguments,
+from source.capacity_settlement.src.capacity_settlement.application.job_args.capacity_settlement_job_args import (
     parse_command_line_arguments,
+    parse_job_arguments,
 )
-from source.electrical_heating.src.electrical_heating.application.entry_points.job_args.environment_variables import (
+from source.capacity_settlement.src.capacity_settlement.application.job_args.environment_variables import (
     EnvironmentVariable,
 )
 
@@ -45,7 +45,6 @@ def sys_argv_from_contract(
 def job_environment_variables() -> dict:
     return {
         EnvironmentVariable.CATALOG_NAME.name: "some_catalog",
-        EnvironmentVariable.TIME_ZONE.name: "some_time_zone",
     }
