+++ source/capacity_settlement/tests/capacity_settlement_tests/application/test_entry_points.py
@@ -1,8 +1,6 @@
 ﻿import pytest
 
-from source.capacity_settlement.src.capacity_settlement.application.entry_points import (
-    entry_point as module,
-)
+from source.capacity_settlement.src.capacity_settlement import entry_point as module
 from source.tests.test_common.entry_points.entry_point_test_util import (
     assert_entry_point_exists,
 )
