--- source/electrical_heating/src/electrical_heating/application/entry_points/__init__.py
@@ -1,1 +1,0 @@
-﻿from .entry_point import execute
