--- source/electrical_heating/src/electrical_heating/application/entry_points/job_args/__init__.py
@@ -1,1 +1,0 @@
-﻿
