+++ source/electrical_heating/tests/electrical_heating_tests/test_entry_points.py
@@ -1,8 +1,6 @@
 ﻿import pytest
 
-from source.electrical_heating.src.electrical_heating.application.entry_points import (
-    entry_point as module,
-)
+from source.electrical_heating.src.electrical_heating import entry_point as module
 from source.tests.test_common.entry_points.entry_point_test_util import (
     assert_entry_point_exists,
 )
